Update simple-slashdot-widen.user.css
Updated .rail-right to flex along the bottom. In order for it to be "beautiful" it did mean hiding some advertising, which until now I'd tried avoiding. Unfortunately the scripting was breaking a more simple and easily understood layout so it had to go. Maybe if web site designers and operators could be better in-tune with these facts it  would help!